Version 3
Mejoras en nomenclatura de tablas sql y directorios.

+++ rb-script/modules/plm/review.del.php
@@ -14,11 +14,15 @@ require_once ABSPATH.'rb-script/class/rb-database.class.php';
 $id = $_GET['id'];
 
 // Borrar de la DB
-$r = $objDataBase->Ejecutar("DELETE FROM $table_name WHERE id=".$id);
-if($r){
-	$arr = ['resultado' => true, 'contenido' => 'Elemento eliminado' ];
+if(G_ACCESOUSUARIO){
+	$r = $objDataBase->Ejecutar("DELETE FROM $table_name WHERE id=".$id);
+	if($r){
+		$arr = ['resultado' => true, 'contenido' => 'Elemento eliminado' ];
+	}else{
+		$arr = ['resultado' => false, 'contenido' => $r['error']];
+	}
 }else{
-	$arr = ['resultado' => false, 'contenido' => $r['error']];
+	$arr = ['resultado' => false, 'contenido' => 'Usuario no inicio sesion'];
 }
 die(json_encode($arr));
 ?>
